third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/element-mutation-expected.txt

This is a testharness.js-based test.
[FAIL] Declarative shadow DOM with shadowrootcustomelementregistry attribute without registry initialized should remain null registry after adoption.
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
